clean up models code
